Base reformatting
