+++ cpmpy/solver_interfaces/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['solver_interface','minizinc_python', 'minizinc_text', 'ortools_python']
+__all__ = ['solver_interface','minizinc_python', 'minizinc_text', 'ortools_python', 'util']
