+++ cpmpy/solver_interfaces/util.py
